Announcement added (without writing into file)
