+++ WorkAndFun/src/main/java/pl/infoshare/Menu.java
@@ -31,8 +31,8 @@ public class Menu {
                 break;
             case 3:
                 System.out.println("Wybrałes 3 - Dodaj ogłoszenie z oferowaną uslugą");
-                    Announcement announcement = new Announcement();
-                    announcement.addAnnouncementOfferService();
+                    AnnouncementService myAnnouncementService = new AnnouncementService();
+                    myAnnouncementService.addAnnouncementOfferService();
                 display();
                 break;
             case 4:
